Fix price type mismatch in homework 2 (change price setter/getter and console scanner input to double)

+++ Applied-Programming-Java/home work 2/Drug.java
@@ -13,12 +13,12 @@ public class Drug {
     }
 
     // Setters and Getters
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 
-    public int getPrice() {
-        return (int) price;
+    public double getPrice() {
+        return price;
     }
 
     public void setType(char type) {
